apps/gpr/requirements.txt
Django==5
psycopg2-binary
djangorestframework==3.15
django-filter
djangorestframework-stubs
drf-yasg
PyJWT
celery[redis]
flower